+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 * Fixed design issue when social tab was not displaying proper form styles
 * Fixed issue when Mail.Ru did not show as active when enabled
 * Fixed issue when some websites had problem with loading comments or widget in sidebar, #327
-
+* Fixed issue when user was not notified about new reply when sent from admin panel, #330
 
 ## 0.0.75 – 05.12.2018